Initial Commit
I am able to retrieve data from the API using Retrofit Library and I
have tried to store this data in form of a graph, so as to make rest of the operations easier later. Due to shortage of time I was just able to retrieve data and show it on the app.